adk_skills/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "adk-skills-demo"
version = "0.1.0"
description = "Google ADK skills demo project with file-based skills."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-adk>=1.25.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
adk-skills-check = "adk_skills_demo_agent.cli:check"
adk-skills-info = "adk_skills_demo_agent.cli:print_info"
adk-skills-smoke-test = "adk_skills_demo_agent.smoke_test:main"

[tool.uv]
dev-dependencies = [
  "ruff>=0.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/adk_skills_demo_agent"]
